feat: cập nhật lại quy tắc đặt tên bảng.
- Danh sách api:
    - Tạo, thay đổi, nhận thông tin sản phẩm.
    - Tạo, thay đổi, xóa, nhận giá sản phẩm.
    - Tạo, thay đổi, nhận kích thước cho sản phẩm.
    - Tạo, thay đổi, nhận danh mục sản phẩm.
    - Tạo, thay đổi, nhận đơn vị tính.
    - Tạo, thay đổi, nhận kích thước sản phẩm.
    - Tạo, thay đổi, nhận thông tin tài khoản.
    - Tạo, thay đổi, nhận thông tin vai trò.
    - Tạo, thay đổi, nhận thông tin người quản lý.
    - Tạo, thay đổi, nhận thông tin nhân viên.
    - Tạo, thay đổi, nhận thông tin khách hàng.
    - Tạo, thay đổi, nhận thông tin khách hàng.
    - Tạo, thay đổi, nhận thông tin chương trình thành viên.
    - Tạo, thay đổi, nhận thông tin khu vực.
    - Tạo, thay đổi, nhận thông tin bàn.
    - Tạo, thay đổi, nhận thông tin cửa hàng.
    - Tạo, thay đổi, nhận thông tin mã khuyến mãi.
    - Tạo, thay đổi, nhận thông tin chương trình khuyến mãi.
    - Tạo đơn hàng, hủy đơn hàng.
    - Tiến hành thanh toán:
        - Thanh toán bằng tiền mặt.

Author: Nguyen Ngoc Phu

+++ backend/api/src/main/java/com/mts/backend/api/customer/request/CreateMembershipTypeRequest.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
@@ -12,8 +14,8 @@ import lombok.NoArgsConstructor;
 public class CreateMembershipTypeRequest {
     private String name;
     private String description;
-    private int requiredPoint;
+    private Integer requiredPoint;
     private String discountUnit;
-    private double discountValue;
+    private BigDecimal discountValue;
     private String validUntil;
 }
